Split out set stream as connected/disconnected state methods
